# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AcadVault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /social/posts:
    get:
      operationId: listPosts
      tags: [social]
      summary: Get the news feed
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [social]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /social/posts/{postId}:
    delete:
      operationId: deletePost
      tags: [social]
      summary: Delete own post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /social/posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [social]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
    post:
      operationId: createComment
      tags: [social]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /social/posts/{postId}/react:
    post:
      operationId: toggleReaction
      tags: [social]
      summary: Like or unlike a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Post:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        imageUrl:
          type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        liked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/PostAuthor"
      required:
        - id
        - content
        - likesCount
        - commentsCount
        - liked
        - createdAt
        - author
    PostAuthor:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImageUrl:
          type: string
      required:
        - id
    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        nextCursor:
          type: integer
      required:
        - posts
    CreatePostBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/PostAuthor"
      required:
        - id
        - postId
        - content
        - createdAt
        - author
    CommentListResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required:
        - comments
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    ReactionResponse:
      type: object
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
      required:
        - liked
        - likesCount
...