# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AcadVault API
 * OpenAPI spec version: 0.1.0
 */
import type { PostAuthor } from "./postAuthor";

export interface Comment {
  id: number;
  postId: number;
  content: string;
  createdAt: Date;
  author: PostAuthor;
}