# lib/api-zod/src/generated/types/commentListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AcadVault API
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from "./comment";

export interface CommentListResponse {
  comments: Comment[];
}